distage-framework: working JS prototype, cleanups required, closes #1290
