+++ distage/distage-core/.jvm/src/main/scala/izumi/distage/modules/platform/CatsIOPlatformDependentSupportModule.scala
@@ -1,9 +1,10 @@
 package izumi.distage.modules.platform
 
-import cats.effect.unsafe.IORuntime
-import izumi.distage.model.definition.{Lifecycle, ModuleDef}
+import cats.effect.unsafe.{IORuntime, IORuntimeConfig, Scheduler}
+import izumi.distage.model.definition.{Id, Lifecycle, ModuleDef}
 import izumi.fundamentals.platform.functional.Identity
 
+import java.util.concurrent.atomic.AtomicReference
 import scala.concurrent.ExecutionContext
 
 private[distage] trait CatsIOPlatformDependentSupportModule extends ModuleDef {
@@ -13,6 +14,20 @@ private[distage] trait CatsIOPlatformDependentSupportModule extends ModuleDef {
         acquire = IORuntime.createDefaultBlockingExecutionContext()
       )(release = _._2.apply()).map(_._1)
   }
+  make[ExecutionContext].named("cpu").from((_: (IORuntime, ExecutionContext))._2)
+
+  // by-name cycles don't work reliably at all, so unfortunately, manual cycle breaking:
+  make[(IORuntime, ExecutionContext)].fromResource {
+    (blockingPool: ExecutionContext @Id("io"), scheduler: Scheduler, ioRuntimeConfig: IORuntimeConfig) =>
+      val cpuRef = new AtomicReference[ExecutionContext](null)
+      lazy val ioRuntime: IORuntime = IORuntime(cpuRef.get(), blockingPool, scheduler, () => (), ioRuntimeConfig)
+      CatsIOPlatformDependentSupportModule.createCPUPool.map {
+        ec =>
+          cpuRef.set(ec)
+          (ioRuntime, ec)
+      }
+  }
+  make[IORuntime].from((_: (IORuntime, ExecutionContext))._1)
 }
 
 object CatsIOPlatformDependentSupportModule {
